# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'LBAW Eventful Web API'
  description: 'Web Resources Specification (A7) for Eventful'
servers:
- url: http://lbaw2122.lbaw.fe.up.pt
  description: Production server
...
paths:
...
  /event:
    get:
      operationId: R202
      summary: 'R202: Search Events'
      description: 'Show search results. Access: PUB'
      tags:
        - 'M02: Events'
      parameters:
      - in: query
        name: search
        description: Text used to search for event titles and event descriptions
        schema:
          type: integer
        required: true
      - in: query
        name: tags
        description: Array with the id of the tags that we want to cross-reference with the resultant events
        schema:
          type: array
          items:
            type: integer
        required: false
      responses:
        '200':
          description: 'Show search results'

    post:
      operationId: R203
      summary: 'R203: Event Create Action'
      description: 'Create an event. Access: USR'
      tags:
        - 'M02: Events'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                image:
                  type: string
                  format: binary
                description:
                  type: string
                location:
                  type: string
                date:
                  type: string
                visibility:
                  type: boolean
                accessability:
                  type: boolean
                capacity:
                  type: integer
                price:
                  type: number
                  format: float
              required:
                - title
                - image
                - description
                - location
                - date
                - visibility
                - accessability
                - capacity
                - price
      responses:
        '302':
          description: 'Redirect after creating event.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful event creation. Redirect to event page.'
                  value: '/event/{event_id}'
                302Failure:
                  description: 'Failed event creation. Redirect to event creation form.'
                  value: '/createEvent'

  # --------------------------------- EVENT ---------------------------------
  /event/{event_id}:
    get:
      operationId: R204
      summary: 'R204: View Event Page'
      description: 'Show event page. Access: PUB'
      tags:
        - 'M02: Events'
      parameters:
      - in: path
        name: event_id
        schema:
          type: integer
        required: true
      responses:
        '200':
          description: 'Show Event Page UI'

    put:
      operationId: R205
      summary: 'R205: Event Update Action'
      description: 'Update an event. Access: HOST'
      tags:
        - 'M02: Events'
      parameters:
      - in: path
        name: event_id
        schema:
          type: integer
        required: true
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                image:
                  type: string
                  format: binary
                description:
                  type: string
                location:
                  type: string
                date:
                  type: string
                visibility:
                  type: boolean
                accessability:
                  type: boolean
                capacity:
                  type: integer
                price:
                  type: number
                  format: float
      responses:
        '302':
          description: 'Redirect after updating event.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful event update. Redirect to event page.'
                  value: '/event/{event_id}'
                302Failure:
                  description: 'Failed event update. Redirect to event update form.'
                  value: '/event/{event_id}/update'

    delete:
      operationId: R206
      summary: 'R206: Delete Event Action'
      description: 'Delete an event. Access: HOST'
      tags:
        - 'M02: Events'
      parameters:
        - in: path
          name: event_id
          schema:
            type: integer
          required: true
      responses:
        '302':
          description: 'Redirect after deleting event.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful event delete. Redirect to homepage.'
                  value: '/'
                302Failure:
                  description: 'Failed event delete. Redirect to event page.'
                  value: '/event/{event_id}'
    
  # --------------------------------- EVENT UPDATE ---------------------------------
...